+++ backend/src/main/java/ku_rum/backend/domain/building/application/BuildingSearchService.java
@@ -7,10 +7,10 @@ import ku_rum.backend.domain.building.domain.repository.BuildingQueryRepository;
 import ku_rum.backend.domain.building.domain.repository.BuildingRepository;
 import ku_rum.backend.domain.building.domain.BuildingAbbrev;
 import ku_rum.backend.domain.category.domain.CategoryDetail;
-import ku_rum.backend.domain.category.response.CategoryCafeteriaDetailResponse;
-import ku_rum.backend.domain.category.response.CategoryDetailResponse;
-import ku_rum.backend.domain.category.response.CategoryKcubeDetailResponse;
-import ku_rum.backend.domain.category.service.CategoryService;
+import ku_rum.backend.domain.category.dto.response.CategoryCafeteriaDetailResponse;
+import ku_rum.backend.domain.category.dto.response.CategoryDetailResponse;
+import ku_rum.backend.domain.category.dto.response.CategoryKcubeDetailResponse;
+import ku_rum.backend.domain.category.application.CategoryService;
 import ku_rum.backend.domain.menu.repository.MenuQueryRepository;
 import ku_rum.backend.global.exception.building.BuildingNotFoundException;
 import ku_rum.backend.global.exception.building.BuildingNotRegisteredException;
