+++ backend/src/main/java/ku_rum/backend/domain/category/application/CategoryService.java
@@ -1,8 +1,8 @@
-package ku_rum.backend.domain.category.service;
+package ku_rum.backend.domain.category.application;
 
 import ku_rum.backend.domain.building.domain.repository.BuildingCategoryQueryRepository;
-import ku_rum.backend.domain.category.repository.CategoryQueryRepository;
-import ku_rum.backend.domain.category.repository.CategoryRepository;
+import ku_rum.backend.domain.category.domain.repository.CategoryQueryRepository;
+import ku_rum.backend.domain.category.domain.repository.CategoryRepository;
 import ku_rum.backend.global.exception.category.CategoryNotExist;
 import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
 import lombok.RequiredArgsConstructor;
