+++ backend/src/test/java/ku_rum/backend/domain/building/service/BuildingCategoryRelatedServiceTest.java
@@ -8,8 +8,8 @@ import ku_rum.backend.domain.building.dto.response.BuildingResponse;
 import ku_rum.backend.domain.building.domain.repository.BuildingQueryRepository;
 import ku_rum.backend.domain.building.domain.repository.BuildingRepository;
 import ku_rum.backend.domain.category.domain.Category;
-import ku_rum.backend.domain.category.response.CategoryCafeteriaDetailResponse;
-import ku_rum.backend.domain.category.response.CategoryKcubeDetailResponse;
+import ku_rum.backend.domain.category.dto.response.CategoryCafeteriaDetailResponse;
+import ku_rum.backend.domain.category.dto.response.CategoryKcubeDetailResponse;
 import ku_rum.backend.domain.menu.domain.Menu;
 import ku_rum.backend.global.exception.category.CategoryNotProvidingDetail;
 import org.junit.jupiter.api.AfterEach;
